Add test case variants that use ClientEJBDirectory.
