+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/ejb/remote/RemoteStatefulEJBFailoverTestCase.java
@@ -25,36 +25,23 @@ package org.jboss.as.test.clustering.cluster.ejb.remote;
 import java.util.PropertyPermission;
 
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.container.test.api.TargetsContainer;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.as.test.clustering.cluster.ClusterAbstractTestCase;
 import org.jboss.as.test.clustering.cluster.ejb.remote.bean.Incrementor;
 import org.jboss.as.test.clustering.cluster.ejb.remote.bean.IncrementorBean;
 import org.jboss.as.test.clustering.cluster.ejb.remote.bean.Result;
 import org.jboss.as.test.clustering.cluster.ejb.remote.bean.StatefulIncrementorBean;
 import org.jboss.as.test.clustering.ejb.EJBDirectory;
 import org.jboss.as.test.clustering.ejb.RemoteEJBDirectory;
-import org.jboss.as.test.shared.TimeoutUtil;
 import org.jboss.as.test.shared.integration.ejb.security.PermissionUtils;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
 
 /**
- * Validates failover behavior of a remotely accessed @Stateful EJB.
  * @author Paul Ferraro
  */
-@RunWith(Arquillian.class)
-@RunAsClient
-public class RemoteStatefulEJBFailoverTestCase extends ClusterAbstractTestCase {
+public class RemoteStatefulEJBFailoverTestCase extends AbstractRemoteStatefulEJBFailoverTestCase {
     private static final String MODULE_NAME = "remote-stateful-ejb-failover-test";
-
-    private static final int COUNT = 20;
-    private static final long CLIENT_TOPOLOGY_UPDATE_WAIT = TimeoutUtil.adjust(5000);
 
     @Deployment(name = DEPLOYMENT_1, managed = false, testable = false)
     @TargetsContainer(CONTAINER_1)
@@ -76,89 +63,7 @@ public class RemoteStatefulEJBFailoverTestCase extends ClusterAbstractTestCase {
                 ;
     }
 
-    @Test
-    public void test() throws Exception {
-        try (EJBDirectory directory = new RemoteEJBDirectory(MODULE_NAME)) {
-            Incrementor bean = directory.lookupStateful(StatefulIncrementorBean.class, Incrementor.class);
-
-            Result<Integer> result = bean.increment();
-            String target = result.getNode();
-            int count = 1;
-
-            Assert.assertEquals(count++, result.getValue().intValue());
-
-            // Bean should retain weak affinity for this node
-            for (int i = 0; i < COUNT; ++i) {
-                result = bean.increment();
-                Assert.assertEquals(count++, result.getValue().intValue());
-                Assert.assertEquals(String.valueOf(i), target, result.getNode());
-            }
-
-            undeploy(this.findDeployment(target));
-
-            Thread.sleep(CLIENT_TOPOLOGY_UPDATE_WAIT);
-
-            result = bean.increment();
-            // Bean should failover to other node
-            String failoverTarget = result.getNode();
-
-            Assert.assertEquals(count++, result.getValue().intValue());
-            Assert.assertNotEquals(target, failoverTarget);
-
-            deploy(this.findDeployment(target));
-
-            // Allow sufficient time for client to receive new topology
-            Thread.sleep(CLIENT_TOPOLOGY_UPDATE_WAIT);
-
-            result = bean.increment();
-            String failbackTarget = result.getNode();
-            Assert.assertEquals(count++, result.getValue().intValue());
-            // Bean should retain weak affinity for this node
-            Assert.assertEquals(failoverTarget, failbackTarget);
-
-            result = bean.increment();
-            // Bean may have acquired new weak affinity
-            target = result.getNode();
-            Assert.assertEquals(count++, result.getValue().intValue());
-
-            // Bean should retain weak affinity for this node
-            for (int i = 0; i < COUNT; ++i) {
-                result = bean.increment();
-                Assert.assertEquals(count++, result.getValue().intValue());
-                Assert.assertEquals(String.valueOf(i), target, result.getNode());
-            }
-
-            stop(this.findContainer(target));
-
-            result = bean.increment();
-            // Bean should failover to other node
-            failoverTarget = result.getNode();
-
-            Assert.assertEquals(count++, result.getValue().intValue());
-            Assert.assertNotEquals(target, failoverTarget);
-
-            start(this.findContainer(target));
-
-            // Allow sufficient time for client to receive new topology
-            Thread.sleep(CLIENT_TOPOLOGY_UPDATE_WAIT);
-
-            result = bean.increment();
-            failbackTarget = result.getNode();
-            Assert.assertEquals(count++, result.getValue().intValue());
-            // Bean should retain weak affinity for this node
-            Assert.assertEquals(failoverTarget, failbackTarget);
-
-            result = bean.increment();
-            // Bean may have acquired new weak affinity
-            target = result.getNode();
-            Assert.assertEquals(count++, result.getValue().intValue());
-
-            // Bean should retain weak affinity for this node
-            for (int i = 0; i < COUNT; ++i) {
-                result = bean.increment();
-                Assert.assertEquals(count++, result.getValue().intValue());
-                Assert.assertEquals(String.valueOf(i), target, result.getNode());
-            }
-        }
+    public RemoteStatefulEJBFailoverTestCase() {
+        super(() -> new RemoteEJBDirectory(MODULE_NAME));
     }
 }
